requirements.txt
redis
aiohttp
waifuim.py
aiosu
jishaku
pynacl
wavelink
discord.py
better-ipc